+++ glutin_egl_sys/src/lib.rs
@@ -26,9 +26,12 @@ pub mod egl {
 
     include!(concat!(env!("OUT_DIR"), "/egl_bindings.rs"));
 
-    // TODO should upstream this.
+    // TODO should upstream these:
+    // EGL_EXT_platform_xcb
     pub const PLATFORM_XCB_EXT: super::EGLenum = 0x31DC;
     pub const PLATFORM_XCB_SCREEN_EXT: super::EGLenum = 0x31DC;
+    // EGL_EXT_device_query_name
+    pub const RENDERER_EXT: super::EGLenum = 0x335F;
 }
 
 pub use self::egl::types::{EGLContext, EGLDisplay};